Update README.md of FunctionItem & function-nodes.md
I fully understood FuntionItem node in the end from checking the source code, but documentation is confusing in several ways. That's because Function and FunctionItem have different styles. Function gives access to the whole items array and the structure is transparently available with json/binary.
In case of FunctionItem. The item is not item, it's item.json. You can't access item.binary cause of this, as it would be intuitive to do so.
At first an update to the documentation to guide new users. It should be also considered that Function and FunctionItem behave the same, e.g. item.json and item.binary should exists. The variable might also called itemjson instead of item.