describe how to start webhook processors correctly
We've followed this step by step guide, but noticed a problem afterwards using cron-triggered jobs. We started the webhook processors like described in this guide as "n8n workers", but instead we had to start it with "n8n webhook". This should be clearly stated in the docs, as it seems like a common pitfall. Maybe my brain-power was too low at that time, too :sweat_smile: