Initial attempt at documenting new Dedupe behavior
This change documents the new deduplication operations recently added to
the Remove Duplicates node. It splits the document by operations type to
better organize around the features specific to different modes of
operation.

Additionally, it splits off a separate "Templates and examples" page to
hold more extensive examples for demonstrating how to use some of the
new operation types.

As part of this change, I've also enabled the `sane_lists` markdown
extension so that numbered steps do not reset when interrupted by other
items like tables and examples. This might have a negative impact on
other pages, so we may have to revert this change after testing.