Extand example to exclude nodes
It's not obvious for a regular user what the node name to use. The current example shows regular nodes names. LLM nodes have different reference. Replacing one of the regular nodes with a reference to the name of an LLM node might help to orient the reader in the right direction.